Added a top layer to References section. Moved dragdrop to the 2nd layer
