+++ midas-author/lps/src/assets/sample-data/nerdm2.json
@@ -121,7 +121,7 @@
             "downloadURL": "http://nigos.nist.gov:8080/nist/sd/32/NIST_SD32_MEDS-I_face.zip",
             "filepath": "NIST_SD32_MEDS-I_face.zip",
             "@type": [
-                "nrdp:AccessPage1",
+                "nrdp:AccessPage",
                 "dcat:Distribution"
             ],
             "@id": "cmps/NIST_SD32_MEDS-I_face.zip",
@@ -140,7 +140,7 @@
             "downloadURL": "http://nigos.nist.gov:8080/nist/sd/32/NIST_SD32_MEDS-I_face.zip",
             "filepath": "NIST_SD32_MEDS-I_face.zip",
             "@type": [
-                "nrdp:AccessPage1",
+                "nrdp:AccessPage",
                 "dcat:Distribution"
             ],
             "@id": "Test01",
