+++ midas-author/lps/src/styles.scss
@@ -28,10 +28,10 @@
     --active-item: #33ccff;
     --data-changed-saved: #dbffdd;
     --data-changed: #fffeee;
-    --editable: #e6f2ff;
+    --editable: #ecf2f9;
     --tag-blue: #c0e6ff;
-    --disabled-grey: #c7c7c7;
-    --disabled-grey-text: #888888;
+    --disabled-grey: #bababa;
+    --disabled-grey-text: #6c6c6c;
 }
 
 .myCustomModalClass .modal-dialog {
