+++ oar-lps/libs/oarlps/src/lib/landing/accesspage/accesspage-edit/accesspage-edit.component.css
@@ -1,0 +1,8 @@
+.label {
+    margin-right: 5px;
+}
+
+td:first-child { 
+    width: 20% ;
+    max-width: 200px;
+}
