+++ oar-lps/libs/oarlps/src/lib/landing/accesspage/accesspage-edit/accesspage-edit.component.html
@@ -1,0 +1,62 @@
+<div class="editable_container" style="width: 100%;padding: 20px;" >
+    <div class="editable_field" style="width: 100%;" [ngStyle]="mdupdsvc.getFieldStyle(fieldName, accessPage.dataChanged)">
+            <table align="center" style="width: 100%;">
+                <!-- Link URL -->
+                <tr>
+                    <td align="right">
+                        <span style="color: rgb(173, 0, 0);margin-right: 6px;">*</span>
+                        <span class="label">Link URL:</span>
+                    </td>
+                    <td>
+                        <span style="width: 100%">
+                            <span style="display:inline-block;width: 100%;">
+                                <input type="text" [(ngModel)]="accessPage.accessURL" style="height:2em;"
+                                (input)="onChange()"/>
+                            </span>
+                        </span>
+                    </td>
+                </tr>
+        
+                <!-- Link Text -->
+                <tr style="width: 100%;">
+                    <td align="right">
+                        <span class="label">Link text:</span>
+                    </td>
+                    <td style="width: 80%">
+                        <span style="width: 100%">
+                            <span style="display:inline-block;width: 100%;">
+                                <input type="text" [(ngModel)]="accessPage.title" style="height:2em;"
+                                (input)="onChange()"/>
+                            </span>
+                        </span>
+                    </td>
+                </tr>
+                <!-- About this link -->
+                <tr>
+                    <td align="right">
+                        <span class="label">About this link:</span>
+                    </td>
+                    <td>
+                        <span style="width: 100%">
+                            <span style="display:inline-block;width: 100%;">
+                                <input type="text" [(ngModel)]="accessPage.description" style="height:2em;"
+                                (input)="onChange()"/>
+                            </span>
+                        </span>
+                    </td>
+                </tr>
+            </table>
+    </div>
+    <div class="editable_field" style="white-space: nowrap;">
+        <button class="edit_button" type="button" (click)="commandOut('saveCurrentChanges')" aria-label="edit button" [disabled]="noURL || !accessPage.dataChanged" ngbTooltip="Save current changes">
+            <i class="faa faa-check"></i>
+        </button>
+
+        <button class="edit_button" type="button"
+            (click)="commandOut('undoCurrentChanges')" aria-label="undo button" ngbTooltip="Undo current changes">
+            <i class="faa faa-remove"></i>
+        </button>
+    </div>
+</div>
+
+
