+++ oar-lps/libs/oarlps/src/lib/landing/accesspage/accesspage-edit/accesspage-edit.component.ts
@@ -1,21 +1,24 @@
 import { Component, OnInit, Input, Output, EventEmitter, SimpleChanges } from '@angular/core';
 import { NerdmRes, NerdmComp, NERDResource } from '../../../nerdm/nerdm';
+import { MetadataUpdateService } from '../../editcontrol/metadataupdate.service';
 
 @Component({
-  selector: 'lib-single-apage',
-  templateUrl: './single-apage.component.html',
-  styleUrls: ['./single-apage.component.css']
+  selector: 'lib-accesspage-edit',
+  templateUrl: './accesspage-edit.component.html',
+  styleUrls: ['../../landing.component.scss', './accesspage-edit.component.css']
 })
-export class SingleApageComponent implements OnInit {
+export class AccesspageEditComponent implements OnInit {
     originalApage: NerdmComp = {} as NerdmComp;
     editBlockStatus: string = 'collapsed';
+    fieldName: string = 'components';
 
     @Input() accessPage: NerdmComp = null;
     @Input() editMode: string = "edit";
     @Input() forceReset: boolean = false;
     @Output() dataChanged: EventEmitter<any> = new EventEmitter();
+    @Output() cmdOutput: EventEmitter<any> = new EventEmitter();
 
-    constructor() { }
+    constructor(public mdupdsvc : MetadataUpdateService) { }
 
     ngOnInit(): void {
         if(this.accessPage) this.originalApage = JSON.parse(JSON.stringify(this.accessPage));
@@ -23,7 +26,10 @@ export class SingleApageComponent implements OnInit {
 
     get isEditing() { return this.editMode=="edit" };
     get isAdding() { return this.editMode=="add" };
-    
+    get noURL() {
+        return !this.accessPage.accessURL || this.accessPage.accessURL.trim() == "";
+    }
+
     ngOnChanges(changes: SimpleChanges): void {
         if(changes.editMode && changes.editMode.currentValue == "normal") {
             this.reset();
@@ -48,4 +54,12 @@ export class SingleApageComponent implements OnInit {
         this.accessPage.dataChanged = true;
         this.dataChanged.emit({"accessPage": this.accessPage, "dataChanged": true});
     }
+
+    /**
+     * Emit command to parent component
+     * @param cmd command
+     */
+    commandOut(cmd: string) {
+        this.cmdOutput.emit({"command": cmd});
+    }    
 }
