+++ oar-lps/libs/oarlps/src/lib/landing/accesspage/accesspage-list/accesspage-list.component.html
@@ -2,7 +2,7 @@
     <!-- <div *ngIf="accessPages && accessPages.length > 0">
         <b>Data and related material can be found at the following locations:</b>
     </div> -->
-    <div *ngIf="mdupdsvc.isEditMode; else nonEdit" style="width: 100%;">
+    <div style="width: 100%;">
         <div class="editable_container" style="width: 100%;">
             <div class="editable_field" style="width: 100%;">
                 <div class="apage-text edit_box" [style.background-color]="getRecordBackgroundColor()">
@@ -41,52 +41,35 @@
                     </div>
                 </div>
             </div>
-            <div class="editable_field">
+            <div class="editable_field" style="white-space: nowrap;">
                 <!-- Add button -->
-                <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" (click)="onAdd()" aria-label="add button" [disabled]="disableEditing || isEditing || isAdding" ngbTooltip="Add access page">
+                <button class="edit_button" type="button" (click)="onAdd()" aria-label="add button" [disabled]="isEditing || isAdding" ngbTooltip="Add access page">
                     <i [class]="addIconClass()"></i>
                 </button>
 
                 <!-- Edit button -->
-                <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" (click)="setMode('edit')" aria-label="edit button" [disabled]="disableEditing || !isNormal" ngbTooltip="Edit access page">
+                <button class="edit_button" type="button" (click)="setMode('edit')" aria-label="edit button" [disabled]="!isNormal || accessPages.length==0" ngbTooltip="Edit access page">
                     <i [class]="editIconClass()"></i>
                 </button>
 
-                
-                <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" [disabled]="!dataChanged && !isEditing"
+                <!-- Restore original value-->
+                <button class="edit_button" type="button" [disabled]="!dataChanged || isEditing || isAdding"
                     (click)="onCancel()" aria-label="undo button" ngbTooltip="Undo changes">
                     <i [class]="undoIconClass()"></i>
                 </button>
 
                 <!-- Hide button - hide edit block -->
-                <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" [disabled]="isAdding || isEditing"
+                <button class="edit_button" type="button" [disabled]="isAdding || isEditing"
                     (click)="hideEditBlock()" aria-label="Hide edit block" ngbTooltip="Hide edit block">
                     <i class="faa faa-remove"></i>
                 </button> 
-
             </div>
         </div>
-        <div style="overflow: hidden; width: 100%;" [@editExpand]="editBlockStatus">  
-            <table style="width: 100%;">
-                <tr  style="width: 100%;">
-                    <td class="edit-block" style="width: calc(100%-30px);">
-                        <lib-single-apage [accessPage]="currentApage" [editMode]="editMode" [forceReset]="forceReset" (dataChanged)="onDataChange($event)"></lib-single-apage> 
-                    </td>
-                    <td style="width: 30px;vertical-align: top;">
-                        <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" (click)="saveCurApage()" aria-label="edit button" [disabled]="!dataChanged" ngbTooltip="Save current changes">
-                            <i class="faa faa-check"></i>
-                        </button>
-        
-                        <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button"
-                            (click)="undoCurApageChanges()" aria-label="undo button" ngbTooltip="Undo current changes">
-                            <i class="faa faa-remove"></i>
-                        </button>
-                    </td>
-                </tr>
-            </table>      
+        <div style="overflow: hidden; width: 100%;" [@editExpand]="editBlockStatus"> 
+            <lib-accesspage-edit [accessPage]="currentApage" [editMode]="editMode" [forceReset]="forceReset" (dataChanged)="onDataChange($event)" (cmdOutput)="onCommandChanged($event)"></lib-accesspage-edit> 
         </div>
     </div>
-    <ng-template #nonEdit>
+    <!-- <ng-template #nonEdit>
         <div *ngIf="accessPages && accessPages.length > 0">
             <div [ngStyle]="{'margin-top': '10px', 'background-color': apage['backcolor']}"
             *ngFor="let apage of accessPages" (mouseover)="apage['backcolor'] = '#f2f2f2'"
@@ -136,5 +119,5 @@
                 </ng-template>
             </div>
         </div>
-    </ng-template>
+    </ng-template> -->
 </div>
