+++ oar-lps/libs/oarlps/src/lib/landing/accesspage/accesspage.component.html
@@ -4,58 +4,61 @@
     </div>
 
     <div class="editable_container" style="padding: 0px;">
-        <div *ngIf="mdupdsvc.isEditMode && (!accessPages || accessPages.length == 0)" [ngStyle]="mdupdsvc.getFieldStyle(fieldName)" style="color: var(--disabled-grey-text); padding-left: 10px;">
-            Click edit (pencil) button to add links to relevant websites including software repositories
-        </div>
+        <div class="editable_field" [ngStyle]="getStyle()" style="padding: 10px; width: calc(100% - 1.65em);" >
+            <div *ngIf="mdupdsvc.isEditMode" style="margin-bottom: .5em;">
+                Add links to relevant websites including software repositories.
+            </div>
 
-        <div class="editable_field" *ngIf="accessPages && accessPages.length > 0" [ngStyle]="mdupdsvc.getFieldStyle(fieldName)" style="padding: 10px; width: calc(100% - 1.65em);" >
-            <div [ngStyle]="{'margin-top': i==0? '0px':'10px', 'background-color': apage['backcolor']}"
-            *ngFor="let apage of accessPages; let i = index" (mouseover)="apage['backcolor'] = '#f2f2f2'"
-            (mouseout)="apage['backcolor'] = getBackgroundColor()">
-                <div class="indent-content"
-                    *ngIf="!apage['@type'].includes('nrdp:RestrictedAccessPage'); else RestrictedAccessPage">
-                    <i class="faa faa-external-link"> <span style="margin-left:0em;">
-                            <span *ngIf="apage['description']; else noAccessTitle">
-                                <a href="{{apage['accessURL']}}" title="{{apage['description']}}" (click)="gaService.gaTrackEvent('homepage', $event, 
-                                                                'Resource title: ' + record.title,
-                                                                apage['accessURL'])">&nbsp;&nbsp;{{apage['title']}}</a>
-                            </span>
-                            <ng-template #noAccessTitle>
-                                <a href="{{apage['accessURL']}}" (click)="gaService.gaTrackEvent('homepage', $event, 
-                                                                'Resource title: ' + record.title,
-                                                                apage['accessURL'])">&nbsp;&nbsp;{{apage['title']}}</a>
-                            </ng-template>
-                        </span></i>
-                    <div *ngIf="apage['description']">
-                        <div style="cursor: pointer;color: #1471AE;" (click)="apage[showDesc] = !apage[showDesc]"><i
-                                class="faa faa-caret-right" *ngIf="!apage[showDesc]"></i><i class="faa faa-caret-down"
-                                *ngIf="apage[showDesc]"></i> About this link
-                        </div>
-                        <div class="description" *ngIf="apage[showDesc]" [@enterAnimation]>{{apage['description']}}
-                        </div>
-                    </div>
-                </div>
-                <ng-template #RestrictedAccessPage>
-                    <div style="padding:10px 20px;">
-                        <a href={{apage.accessURL}} target="_blank">
-                            <button style="margin-top: -6px;" type="button" class="btn btn-primary py-0"
-                                data-toggle="tooltip" data-placement="top" title="{{apage['title']}}"
-                                (click)="googleAnalytics(apage['accessURL'], $event, 'Download Data')"><i
-                                    class="faa faa-external-link button_icon"></i>Download Data</button>
-                        </a>
-
+            <div *ngIf="accessPages && accessPages.length > 0">
+                <div [ngStyle]="{'margin-top': i==0? '0px':'10px', 'background-color': apage['backcolor']}"
+                *ngFor="let apage of accessPages; let i = index" (mouseover)="apage['backcolor'] = '#f2f2f2'"
+                (mouseout)="apage['backcolor'] = 'transparent'">
+                    <div class="indent-content"
+                        *ngIf="!apage['@type'].includes('nrdp:RestrictedAccessPage'); else RestrictedAccessPage">
+                        <i class="faa faa-external-link"> <span style="margin-left:0em;">
+                                <span *ngIf="apage['description']; else noAccessTitle">
+                                    <a href="{{apage['accessURL']}}" title="{{apage['description']}}" (click)="gaService.gaTrackEvent('homepage', $event, 
+                                                                    'Resource title: ' + record.title,
+                                                                    apage['accessURL'])">&nbsp;&nbsp;{{apage['title']}}</a>
+                                </span>
+                                <ng-template #noAccessTitle>
+                                    <a href="{{apage['accessURL']}}" (click)="gaService.gaTrackEvent('homepage', $event, 
+                                                                    'Resource title: ' + record.title,
+                                                                    apage['accessURL'])">&nbsp;&nbsp;{{apage['title']}}</a>
+                                </ng-template>
+                            </span></i>
                         <div *ngIf="apage['description']">
                             <div style="cursor: pointer;color: #1471AE;" (click)="apage[showDesc] = !apage[showDesc]"><i
-                                    class="faa faa-caret-right" *ngIf="!apage[showDesc]"></i><i
-                                    class="faa faa-caret-down" *ngIf="apage[showDesc]"></i> About downloading restricted
-                                public data
+                                    class="faa faa-caret-right" *ngIf="!apage[showDesc]"></i><i class="faa faa-caret-down"
+                                    *ngIf="apage[showDesc]"></i> About this link
                             </div>
                             <div class="description" *ngIf="apage[showDesc]" [@enterAnimation]>{{apage['description']}}
                             </div>
                         </div>
                     </div>
-                </ng-template>
+                    <ng-template #RestrictedAccessPage>
+                        <div style="padding:10px 20px;">
+                            <a href={{apage.accessURL}} target="_blank">
+                                <button style="margin-top: -6px;" type="button" class="btn btn-primary py-0"
+                                    data-toggle="tooltip" data-placement="top" title="{{apage['title']}}"
+                                    (click)="googleAnalytics(apage['accessURL'], $event, 'Download Data')"><i
+                                        class="faa faa-external-link button_icon"></i>Download Data</button>
+                            </a>
+    
+                            <div *ngIf="apage['description']">
+                                <div style="cursor: pointer;color: #1471AE;" (click)="apage[showDesc] = !apage[showDesc]"><i
+                                        class="faa faa-caret-right" *ngIf="!apage[showDesc]"></i><i
+                                        class="faa faa-caret-down" *ngIf="apage[showDesc]"></i> About downloading restricted
+                                    public data
+                                </div>
+                                <div class="description" *ngIf="apage[showDesc]" [@enterAnimation]>{{apage['description']}}
+                                </div>
+                            </div>
+                        </div>
+                    </ng-template>
+                </div>
             </div>
+
         </div>
 
         <!-- Control buttons -->
