--- oar-lps/libs/oarlps/src/lib/landing/accesspage/single-apage/single-apage.component.html
@@ -1,46 +1,0 @@
-<div style="width: 100%; padding: 0px 20px 20px 20px;">
-    <table align="center" style="width: 100%;">
-        <!-- Title -->
-        <tr *ngIf="isEditing || isAdding" style="width: 100%;">
-            <td align="right" style="width: 20% !important;">
-                <span>Title:</span>
-            </td>
-            <td style="width: 80%">
-                <span style="width: 100%">
-                    <span style="display:inline-block;width: 100%;">
-                        <input type="text" [(ngModel)]="accessPage.title" style="height:2em;"
-                        (input)="onChange(true)"/>
-                    </span>
-                </span>
-            </td>
-        </tr>
-        <!-- Description -->
-        <tr *ngIf="isEditing || isAdding">
-            <td align="right">
-                <span>Description:</span>
-            </td>
-            <td>
-                <span style="width: 100%">
-                    <span style="display:inline-block;width: 100%;">
-                        <input type="text" [(ngModel)]="accessPage.description" style="height:2em;"
-                        (input)="onChange(true)"/>
-                    </span>
-                </span>
-            </td>
-        </tr>
-        <!-- Access URL -->
-        <tr *ngIf="isEditing || isAdding">
-            <td align="right">
-                <span>Access URL:</span>
-            </td>
-            <td>
-                <span style="width: 100%">
-                    <span style="display:inline-block;width: 100%;">
-                        <input type="text" [(ngModel)]="accessPage.accessURL" style="height:2em;"
-                        (input)="onChange(true)"/>
-                    </span>
-                </span>
-            </td>
-        </tr>
-    </table>
-</div>
