+++ oar-lps/libs/oarlps/src/lib/landing/references/ref-list/ref-list.component.html
@@ -1,0 +1,95 @@
+<div class="references">
+    <!-- <div class="section-header"><h3><b>References</b></h3></div> -->
+    <div class="editable_container"  style="width: 100%;">
+        <div class="editable_field" style="width: 100%;">
+            <!-- Drag and drop reference list-->
+            <div class="reference-text edit_box" [style.background-color]="getRecordBackgroundColor()">
+                <!-- <div style="margin-left: 10px;"><i>Use this area to add references to papers and datasets</i></div> -->
+                <div #dropListContainer class="example-container" cdkDropListGroup>
+                    <div style="width: 100%;" 
+                    *ngFor="let reference of record['references']; let i = index"
+                    cdkDropList
+                    [cdkDropListData]="i"
+                    >
+                        <div 
+                            cdkDrag
+                            [cdkDragData]="i"
+                            (cdkDragEntered)="dragEntered($event)"
+                            (cdkDragMoved)="dragMoved($event)"
+                            (cdkDragDropped)="dragDropped($event)"
+                            [cdkDragDisabled]="isAdding || isEditing"
+                            class="ref-row"
+                        >
+                            <div (click)="selectRef(i)" [ngStyle]="getActiveItemStyle(i)" style="width: 100%;"  [style.background-color]="getBackgroundColor(i)">
+                                <lib-text-edit 
+                                    [textField]="reference.citation" 
+                                    [dragDropIcon]="record['references'].length==1?false:true" 
+                                    [editButton]="false" 
+                                    [deleteButton]="true" 
+                                    [restoreButton]="true" 
+                                    [disableControl]="isAdding || isEditing" 
+                                    [placeHolderText]="placeholder" 
+                                    [isNew]="reference.isNew" 
+                                    [dataChanged]="mdupdsvc.fieldUpdated(fieldName, record[fieldName][i]['@id'])" 
+                                    (command_out)="onReferenceCommand($event, i)">
+                                </lib-text-edit>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <!-- Control buttons for drag and drop sectioon -->
+        <div class="editable_field" [style.width.px]="getControlBoxWidth()">
+            <!-- Add button -->
+            <button class="edit_button" type="button" (click)="onAdd()" aria-label="add button" [disabled]="isEditing || isAdding" ngbTooltip="Add references">
+                <i [class]="addIconClass()"></i>
+            </button>
+
+            <!-- Edit button -->
+            <button class="edit_button" type="button" (click)="onEdit()" aria-label="edit button" [disabled]="!hasDisplayableReferences() || !isNormal" ngbTooltip="Edit references">
+                <i [class]="editIconClass()"></i>
+            </button>
+
+            <!-- Undo all changes button -->
+            <button class="edit_button" type="button" [disabled]="!dataChanged || isEditing || isAdding"
+                (click)="undoChanges()" aria-label="undo button" ngbTooltip="Undo all reference changes">
+                <i [class]="undoIconClass()"></i>
+            </button>
+
+            <!-- Hide button - hide edit block -->
+            <button class="edit_button" type="button" [disabled]="isAdding || isEditing || isAdding"
+                (click)="hideEditBlock()" aria-label="Hide edit block" ngbTooltip="Hide edit block">
+                <i class="faa faa-remove"></i>
+            </button> 
+
+        </div>
+    </div>
+    <!-- This is editing area. Will open up in add or edit mode -->
+    <div style="overflow: hidden; width: 100%;" [@editExpand]="editBlockStatus">  
+        <lib-ref-edit [ref]="currentRef" [editMode]="editMode" [forceReset]="forceReset" (dataChanged)="onReferenceChange($event)"  (cmdOutput)="onCommandChanged($event)"></lib-ref-edit>
+        
+        <!-- <table style="width: 100%;">
+            <tr  style="width: 100%;">
+
+                <td class="edit-block" style="width: calc(100%-30px);">
+                    <lib-ref-edit [ref]="currentRef" [editMode]="editMode" [forceReset]="forceReset" (dataChanged)="onReferenceChange($event)"  (cmdOutput)="onCommandChanged($event)"></lib-ref-edit>
+                </td>
+
+
+                <td style="width: 30px;vertical-align: top;">
+
+                    <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button" (click)="saveCurRef()" aria-label="edit button" [disabled]="!dataChanged" ngbTooltip="Save current changes">
+                        <i class="faa faa-check"></i>
+                    </button>
+    
+
+                    <button class="edit_button" *ngIf="mdupdsvc.isEditMode" type="button"
+                        (click)="undoCurRefChanges()" aria-label="undo button" ngbTooltip="Undo current changes">
+                        <i class="faa faa-remove"></i>
+                    </button>
+                </td>
+            </tr>
+        </table>       -->
+    </div>
+</div>
