+++ oar-lps/libs/oarlps/src/lib/landing/references/references.module.ts
@@ -5,7 +5,7 @@ import { ToolbarModule } from 'primeng/toolbar';
 import { ToastrModule } from 'ngx-toastr';
 import { ButtonModule } from 'primeng/button';
 import { FormsModule } from '@angular/forms';
-import { SingleRefComponent } from './single-ref/single-ref.component';
+import { RefEditComponent } from './ref-edit/ref-edit.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { DropdownModule } from 'primeng/dropdown';
@@ -14,9 +14,10 @@ import { CollapseModule } from '../collapseDirective/collapse.module';
 import { TextEditModule } from '../../text-edit/text-edit.module';
 import { RefAuthorComponent } from './ref-author/ref-author.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { RefListComponent } from './ref-list/ref-list.component';
 
 @NgModule({
-  declarations: [ReferencesComponent, SingleRefComponent, RefAuthorComponent],
+  declarations: [ReferencesComponent, RefEditComponent, RefAuthorComponent, RefListComponent],
   imports: [
     CommonModule,
     ToolbarModule,
@@ -33,10 +34,10 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
   providers: [
     LandingpageService
   ],
-  exports: [ReferencesComponent, RefAuthorComponent]
+  exports: [ReferencesComponent, RefAuthorComponent, RefListComponent]
 })
 export class ReferencesModule { }
 
 export {
-    ReferencesComponent, RefAuthorComponent
+    ReferencesComponent, RefAuthorComponent, RefListComponent
 }
