+++ oar-lps/libs/oarlps/src/lib/landing/visithome/visithome-edit/visithome-edit.component.html
@@ -5,7 +5,7 @@
         <td class="edit-block" style="width: calc(100%-30px);" [style.background-color]="backgroundColor">
             <div style="width: 100%;text-align: left;padding: 10px;">
                 <div style="width: 100%;">
-                    Enter a new url if you would like to change the current homepage location.
+                    Leave this blank to use this page as your homepage. If you have your own custom homepage, enter its url here.
                 </div>
                 <table width="100%" align="left">
                     <tr>
